Add dittochance tag to Pokemon XML file and change genders percentage from uint8_t to float

+++ src/enums.h
@@ -497,8 +497,8 @@ struct Outfit_t {
 };
 
 struct Gender_t {
-	uint8_t male = 0;
-	uint8_t female = 0;
+	float male = 0;
+	float female = 0;
 };
 
 struct LightInfo {
